refactor: move evaluateFormula to helpers
