+++ src/helpers/index.js
@@ -1,1 +1,2 @@
 export * from './appendOperator'
+export * from './evaluateFormula'
